unpack-bootstrap-tools.sh: use patchelf to undo nuke-refs and nothing else
The current unpack-bootstrap-tools.sh wields patchelf selectively,
guided by unexplained criteria.  Specifically, it applies

patchelf --set-interpreter to:

  $out/bin/* $out/libexec/gcc/*/*/* (except */liblto*)

and patchelf --set-interpreter --set-rpath to:

  $out/lib/librt-*.so $out/lib/libpcre*

The provenance of these lists is unclear.  Some libraries which never
had an rpath to begin with (like librt.so) are given an rpath by this
script, while other libraries (like libbz2.so) are left with dangling
/nix/store/eeee...'s.

This makes it very difficult to troubleshoot the behavior of the
bootstrap-tools unpacker when porting to new platforms, where it is
likely the case that patchelf is being debugged concurrently with
nixpkgs.

This commit changes unpack-bootstrap-tools.sh to use patchelf *only*
to replace /nix/store/eeee... references created by nuke-refs, and to
replace *all* such references except where an exception is clearly
described and explained.  Currently there are two such exceptions:

  1. libgcc_s.so is exempted because its rpath leaks into the final
     stdenv, so we cannot patchelf its rpath without creating a
     fordbidden stdenv-final->bootstrap-tools requisite.

  2. libpthread.so and libc.so are exempted from --set-interpreter
     because the ELF interpreter of these files does not matter (they
     are libraries, not executable programs, and unlike ld.so they are
     not meant to play both roles), and because an unresolved bug in
     patchelf causes corruption if it is used to set the interpreter
     here:

     https://github.com/NixOS/patchelf/issues/368

Making sure to patchelf away all of the nuke-refs'd paths will make
the bootstrapping process more robust and less fragile in general.  In
particular, it will prevent situations like #169746 where the only
thing preventing a bug from manifesting was the fact that the
arbitrary choice of which libraries to patchelf had been made in a
particular way.  Situations like that make things more fragile,
because now the particular choice which used to be arbitrary has
become mandatory, and neither that requirement nor the reason for it
is documented anywhere.

This commit requires #169746 (which is included in this PR) in order
to pass CI.