haskellPackages.reflection: disable tests for GHC 9.0.2 everywhere
Thanks to Ryan Scott's investigation it became clear this crash can also
happen on Linux and thus seems to be a general issue in GHC 9.0.2 (only,
luckily). We'll disable the test suite everywhere for now, but the
prospect of this crash / incorrect runtime result is a bit ominous.
Refer to the [ghc-issue] for details.

[ghc-issue]: https://gitlab.haskell.org/ghc/ghc/-/issues/21141