nghttp2: 1.43.0 -> 1.47.0
Refactor of the feature switches as there was some invalid combinations.
{c-ares,libev,openssl,zlib} are all strictly needed if we want to build
the applications but are unnecessary when building the lib alone.

Accordingly the overrides  in `fetchurl` and `pythonPackages.nghttp2`
are updated to use `enableApp = false` as they only want libnghttp2.
fetchurl bootstrapping just got lighter. :)

Also enable the checks whenever possible. To run them we need to set
TZDATA environment variable and the `CUnit` tool.
Those tests are made optionals as we don't want to scope `cunit` and
`tzdata` when bootstrapping `fetchurl`.

Also install the Bash completions for h2load, nghttp, nghttpd and nghttpx.

Finally, add `c0bw3b` as maintainer.