Fix parsed static desktop items by allowing "true" | "false"
I was trying to update the mailreader Himalaya which uses a *.desktop
file in their Git repository `cp`’d over in the build step. This
*should* be supported however, the new code from the GitHub merge
request #161374, the desktop file fails on `Terminal=true` saying the
value must be a boolean or none. This makes sense, but the way this
configuration file is parsed, `true` is not considered a boolean. This
is interesting because the `boolOrNullToString` checks for boolean type
then immediately casts to string. My quick & dirty solution is to check
if the string is specifically `true` or `false` and if so, allow it.

Additionally `nixpkgs-fmt` was ran, because these are the `nixpkgs`.