buildDotnetModule: allow fetching nuget dependencies at compile time
This allows user to specify a hash for the dependencies using
the "nugetSha256" attribute, instead of having to manually generate
a lockfile.

This is done by generating a lockfile with `dotnet restore`, and then
parsing the requested version ranges to see if anything is floating.

Afterwards we generate a nix-based lockfile containing hashes and stable
downloads for each dependency, which we can IFD. The checksum of this
lockfile is specified with "nugetSha256".

We want to use the checksum of the nix-based lockfile instead of hashing
the entire nuget source so that we can independantly fetch all
dependencies, and re-use them across derivations.