nixos/tests: fix flaky cntr test
The cntr sometimes hangs until the 10-hour hydra limit. This behaviour
appears to be an edge-case related to the type of TTY in which the cntr
command runs during test execution. We can work around this by running
the command as a background job.

I additionally added a wait_for_open_port to fix nondeterministic test
failures I observed after fixing the hanging issue.