source-{sans,serif}-pro: reintroduce older font versions
Recently, Source Sans/Serif Pro fonts got renamed. The "Pro" suffix
was replaced with the major version of the font name.

In the case of Source Serif, the change was announced in
https://github.com/adobe-fonts/source-serif/issues/77 and besides the
name font metric have been changed too. Therefore, documents using the
old "Pro" version might look differently when the font is changed to
the new "4" version.

In the case of Source Sans, the metrics seem not to have been
changed (https://github.com/adobe-fonts/source-sans/issues/192), but
this comment (https://github.com/adobe-fonts/source-sans/issues/192#issuecomment-790143280)
suggests that it might happen in the future.

Due to the above, it makes sense to keep both old "Pro" and
new (numbered) fonts as separate packages. This commit reintroduced
the old "Pro" versions.

It seems that Arch Linux is also distributing both versions:
https://github.com/archlinux/svntogit-packages/blob/0d70461e5a26cde02cadfda64078eb93192c6caa/trunk/PKGBUILD

Closes #157972