zfs: Remove top-level `with` from expression
Using `with` in such a high up scope makes it very easy to
accidentally forget to update function arguments and introduce typos
because of the lack of any editor support and static analysis.

It's better to explicitly inherit the small number of things we need
from `lib` into scope so that editors can know of scoped variables.