wpa_supplicant: rename withDbus to dbusSupport
wpa_supplicant/default.nix currently uses the option "withDbus" to
decide whether or not to compile with dbus support.  It is the only
nix package that uses this choice of name.  Most other packages use
dbusSupport instead.

Let's rename this option to dbusSupport, so that users desiring to
compile without dbus can set the option once in nixpkgs.conf and have
all packages understand that directive.