recoll: prevent Makefile from seeing /etc/debian_version
Recoll's Makefile assumes the installation should be formatted for
Debian if the file /etc/debian_version exists.  This test will succeed
if nixpkgs is used on Debian without the sandbox (which is not
supported on all architectures).

Let's substituteInPlace the Makfile, replacing a test for the
existence of /etc/debian_version with a call to the "false" binary.