evince: add withGnome option (defaults to true on Linux)
This commit adds a withGnome option (defaults to true on Linux) to the
evince derivation, allowing to compile evince independently of the
rest of gnome.  The withGnome option name is also used by eight other
derivations in nixpkgs.

Although evince is part of the gnome project, its build system
supports compiling it independently from the rest of gnome simply by
disabling thumbnail caching.

Evince is currently the only PDF viewer with form-filling support that
doesn't depend on QT.  Nowadays a lot of unavoidable things like tax
forms require a form-filling PDF viewer.  Since people don't get much
of a choice anymore about having to use a piece of software with this
feature, let's at least allow them to do the job with the smallest
possible set of dependencies.

Checked with nixpkgs-hammer, which did not report any new warnings
introduced by this commit.