evince: respect global dbusSupport flag
This commit allows to compile evince without dbus (-Ddbus=false) if
the dbusSupport attribute is set to false.  This can be done globally
in nixpkgs.conf, and is understood by other packages.  See also PR

For reasons which remain unclear, the process of building the
development documentation (devdocs) fails when -Ddbus=false is passed
to meson.  Let's simply not build the devdocs in this case.

Evince is currently the only PDF viewer with form-filling support that
doesn't depend on QT.  Nowadays a lot of unavoidable things like tax
forms require a form-filling PDF viewer.  Since people don't get much
of a choice anymore about having to use a piece of software with this
feature, let's at least allow them to do the job with the smallest
possible set of dependencies.

Checked with nixpkgs-hammer, which did not report any new warnings
introduced by this commit.