pleroma: Potentially wrap binaries with RELEASE_COOKIE
Pleroma_ctl and pleroma depend on mix_release to run which itself
depends on a RELEASE_COOKIE. It'll fail to run without such a cookie.

Allowing the user to wrap this binary with a RELEASE_COOKIE. We don't
set any by default, meaning this diff is no-op for the existing
deployments relying on the pleroma package but not on the NixOS module.