nixos/pleroma: regenerate empty release cookie files
Since b9cfbcafdf0ca9573de1cdc06137c020e70e44a8, the lack of hexdump in
the closure lead to the generation of empty cookie files. This empty
cookie file is making pleroma to crash at startup now we correctly
read it.

We introduce a migration forcing these empty cookies to be
re-generated to something not empty.