gdb: grab upstream patch to fix w/musl
Fixes instances of PAGESIZE/PAGE_SIZE that conflict
with header-provided definitions.

I was concerned about stability (IIRC version in these can change?)
but already risking that with the gcc-12 patch, so this is probably
okay.

Stale/unstable URL's are especially important to avoid on less
commonly used combinations such as w/musl, but also don't
want to bloat the tree :).

Risk is low, as this will be included in next update.