gopls: 0.8.0 -> 0.8.1
Update gopls.

This also changes the version of go used to build it to go1.18. This is
required for gopls to work with go1.18 code.

It's still backwards compatible with go1.17 code, so building with 1.18
should allow us to support all supported go versions (1.17+1.18)
correctly, while building with 1.17 does not.

Once #164320 merges, this should switch back to buildGoModule.