linuxPackages.rtl8821cu: 2021-10-21 -> 2022-03-08
Like with other morrownr modules, the 8821cu repo now links to
the 8821cu-20210118 repository as the place to get the driver code.  I
kept the homepage as 8821cu though, because in future if a new version
of the driver is created, that's where the link will be posted.

Fixes the build with Linux 5.17.