haskellPackages.large-hashable: 0.1.0.4 -> unstable-2021-11-01
master has accumulated not a lot, but significant changes, requiring to
apply a lot of intermediate patches to make newer fixes apply. Since
nothing in the package set depends on large-hashable and the changes
since 0.1.0.4 are (mostly?) internal, I didn't bother with backporting
the patch for aeson 2.0 to apply on top of the hackage tarball.