cgit-pink: init at 1.3.0
cgit-pink is a fork of cgit that aims to be better maintained, because
cgit doesn't get a lot of attention any more, and almost every patch
sent in the last couple of years has been ignored.

The build system is exactly the same as cgit's, so I've created a
common cgit builder function.  This way, improvements to cgit
packaging (I've already noticed a couple to make!) can be shared
between both packages.  If the build systems diverge, we can separate
them out in future.