nixos-rebuild: Start writing simple instrumentation & tests
We want to create a test suite for the branches of nixos-rebuild, so
the first step is making it print some decipherable json insetad of
arbitrary text strings.

Then we can use a python script to decode the json lines, and check
they match what we expect.

Currently the nixos-rebuild is still piped into the test script, it
should probably be called from the test … on the other hand, we need
to run the tests in VM setups anyway if we want to test anything that
is interesting.