pyudev: depend on udev instead of systemd
Pyudev depends only on a libudev implementation, rather than on
systemd-udevd specifically or on systemd generally.  Let's adjust the
dependency list to make it more specific.  In particular, this allows
pyudev to build and run with the overlay which substitutes eudev for
udev.

Pyudev is used by the auxiliary trackpad calibration tools included as
part of libinput.  Because an enormous number of packages depend on
libinput, this allows a very large number of packages to be used
without systemd.  Note that libinput.so does not depend on pyudev or
systemd -- only the trackpad calibration utilities bundled with
libudev use pyudev.