pakcs: build using GHC 8.10.7
Possibly updating the compiler will make it possible to build it with
GHC 9.0.2 out of the box or patching it to work with it would be
trivial (only errors so far are whitespace-related), but I couldn't be
bothered at this time. Sadly the maintainer of the package has
disappeared since adding it.