gobject-introspection: fix cross-compilation
Currently pkgsCross.aarch64-multiplatform.gobject-introspection does
not build due to breakage in (a) generating the docs and (b)
generating the introspection data.

There really isn't proper support in gobject-introspection for
cross-compiles with -Dintrospection=true, just a collection of hacks
in meson_options.txt that are rumored to be useful somehow.  I have
not been able to find any reports of people actually getting this to
work.

Let's disable the docs and introspection data for cross-compiles,
since a build without docs or introspection is still very useful, and
definitely more useful than no build at all.