Add block size to dd command
Without configuring the block size a default of 512 bytes is used, which can
slow down the transfer speed massively.

In a test I've done with a semi-decent USB stick, I only get a transfer speed
of around 180 KB/sec when not specifying the block size but see 27 MB/sec when
setting the block size to 1 MB. This makes the transfer of the minimal
installation ISO take half a minute instead of an hour.