Don't reexec `nixos-rebuild --flake` when host and target systems differ
nixos-rebuild does not work when when building a flake and the host
and target systems are not the same. For example, running the
following on a Darwin system:

    darwin# nixos-rebuild switch --flake ... --build-host linuxhost --target-host linuxhost --config ...

fails with the following error:

    error: a 'x86_64-linux' with features {} is required to build
    '/nix/store/yq1xkqyhb6qh9z4j1bw7b9kdfb0qai2z-nixos-rebuild.drv',
    but I am a 'aarch64-darwin' with features {benchmark,
    big-parallel, nixos-test}

The reason is that nixos-rebuild tries to reexec itself with the newer
version being built. In the example above, it would require
cross-compiling bash and many other tools for Darwin, which would be
problematic for many reasons. Therefore, we simply do not reexec
nixos-rebuild in this case.

Fixes #166499