pcre: fix hardwired reference to /usr/bin/file in configure
The configure script that ships with pcre expects to be able to use
"file", via the hardwired path "/usr/bin/file".  When cross-compiling
in a sandboxed build this will cause it to mis-detect the ability to
create shared libraries.

Let's fix this, and perform the substitution only for cross compiles,
in order to avoid a mass-rebuild.

Currently this problem is preventing Hydra from generating the
bootstrap tarball for mips64el-linux-gnuabi64.